Fix parsing version of BIRD2
In BIRD codes, the label generated from 'git tag' is preferred than fixed version.
https://gitlab.nic.cz/labs/bird/-/blob/c26c6bc2d78a2fe76f27dcc9fbb5afc95c3a7626/sysdep/config.h#L13-17

If we apply our own pathes which are not merged to original BIRD's code to tagged branch, for example 'v2.0.7',
then BIRD_VERSION will be like 'v2.0.7-1-gabcdedfh'.
https://gitlab.nic.cz/labs/bird/-/blob/c26c6bc2d78a2fe76f27dcc9fbb5afc95c3a7626/Makefile.in#L43-46

This patch make birdwatcher to tolerate the expression and enable it to parse BIRD version correctly.